Issue #49 - Implement shared memory mirroring.
Move the mirror_list from unimplmented into its own src/rest
module where we can implement the actual functionality.
